requirements.txt
aio_pika
PyQt5
qasync
soundfile
sounddevice
samplerate
requests
eyeD3

google-auth
